services/auth_service/pyproject.toml
[project]
name = "auth_service"
version = "0.1.0"
description = "auth service"
authors = [
    {name = "taua-almeida",email = "tauan96@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "psycopg[binary] (>=3.2.4,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ruff = "^0.9.3"

